mov value set printing to a separate file
