+++ telamon-gen/src/print/store.rs
@@ -2,6 +2,7 @@
 use ir;
 use ir::SetRef;
 use itertools::Itertools;
+use print::value_set;
 use print::ast::{self, Variable, LoopNest};
 use print::choice::Ast as ChoiceAst;
 use print::choice::CounterValue;
@@ -84,7 +85,7 @@ pub fn incr_iterators<'a>(ir_desc: &'a ir::IrDesc) -> Vec<IncrIterator<'a>> {
                 let mut loop_nest = ast::LoopNest::new(counter_loops, ctx, conflicts, true);
                 conflicts.extend(PartialIterator::new_objs_conflicts(ir_desc, set));
                 loop_nest.extend(incr_loops, ctx, conflicts, false);
-                let incr_condition = RcStr::new(ast::value_set(incr_condition, ctx));
+                let incr_condition = RcStr::new(value_set::print(incr_condition, ctx));
                 let incr = ast::ChoiceInstance::new(incr, ctx);
                 let set_ast = ast::SetDef::new(set.def());
                 let incr_iterator = IncrIterator {
